Rename ScenarioIsolating tester to EnvironmentIsolating & move it to Testwork
