+++ src/Behat/Testwork/Environment/Tester/EnvironmentIsolatingTester.php
@@ -1,16 +1,16 @@
 <?php
 
 /*
- * This file is part of the Behat.
+ * This file is part of the Behat Testwork.
  * (c) Konstantin Kudryashov <ever.zet@gmail.com>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
 
-namespace Behat\Behat\Tester\Gherkin;
+namespace Behat\Testwork\Environment\Tester;
 
-use Behat\Behat\Tester\Context\ScenarioContext;
+use Behat\Testwork\Environment\Context\EnvironmentIsolatingContext;
 use Behat\Testwork\Environment\EnvironmentManager;
 use Behat\Testwork\Tester\Context\TestContext;
 use Behat\Testwork\Tester\Control\RunControl;
@@ -18,11 +18,11 @@ use Behat\Testwork\Tester\Exception\WrongContextException;
 use Behat\Testwork\Tester\Tester;
 
 /**
- * Isolates scenario environment before passing it to the decorated tester.
+ * Isolates context environment before passing it to the decorated tester.
  *
  * @author Konstantin Kudryashov <ever.zet@gmail.com>
  */
-final class ScenarioIsolatingTester implements Tester
+final class EnvironmentIsolatingTester implements Tester
 {
     /**
      * @var Tester
@@ -61,19 +61,19 @@ final class ScenarioIsolatingTester implements Tester
      *
      * @param TestContext $context
      *
-     * @return ScenarioContext
+     * @return EnvironmentIsolatingContext
      *
      * @throws WrongContextException
      */
     private function castContext(TestContext $context)
     {
-        if ($context instanceof ScenarioContext) {
+        if ($context instanceof EnvironmentIsolatingContext) {
             return $context;
         }
 
         throw new WrongContextException(
             sprintf(
-                'ScenarioTester tests instances of ScenarioContext only, but %s given.',
+                'EnvironmentIsolatingTester tests instances of EnvironmentIsolatingContext only, but %s given.',
                 get_class($context)
             ), $context
         );
